requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
joblib==1.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
